requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
google-auth==2.29.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==29.0.0
MarkupSafe==2.1.5
mini-racer==0.12.2
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
zipp==3.19.0
